Write Data to Text File

This function takes a list of data elements (matrices, vectors, or other values) and writes them to a specified text file. Each data element in the list is preceded by a comment line indicating its name.

Usage

write_dat(output_file = "runs/dat/newfile.dat", indata = in_data)

Arguments

output_file

A character string specifying the name of the output text file. Default is "output.txt".

indata

A list containing the data elements to be written to the file. Each element can be a matrix, a vector, or any other value. Default is in_data.

Value

This function doesn't return a value; it writes to the specified output file.

Examples

# Assuming 'data_list' is a list of data elements
# write_dat(output_file = "sample_output.txt", indata = data_list)